test: add runtime shim tests
