+++ test/index.test.mjs
@@ -1,0 +1,2 @@
+import './control-bus.test.mjs';
+import './shims.test.mjs';
